Stop warning about file copies with unmatched sources
We have almost no choice but to represent these in Git as
modifications anyway.  Warnings about unmatched sources on directory
copies are actually useful in completing the rule list, and they are
hard to pick out from all the noise.

Actually I think matching the source may only be a concern for
conversions like the ones done by KDE, which aren't trying to account
for the whole repository.  Perhaps and we should perhaps stop NULLing
out the source path when it isn't matched.