Simply ignore deletion of the master branch.
Currently it prevents us from pushing to Github or BitBucket, and
really it's meaningless.  It only happens when a sub-project is
deleted from from the monolithic SVN project, but that doesn't mean
the sub-project's final commits should become un-referenced in its own
repository!  At most, we should drop the sub-project as a submodule.