ruleset.cpp/.hpp: Remember which branch rules are (potentially) shared among repositories
Only branch rules from abstract repositories can be shared.  We are
going to use this as a heuristic to name branches in the super-module
repository.  A branch that is definitely not shared will be prefixed
with its repository name, to avoid colliding with other
similarly-named branches there.  So if there are two submodules "foo"
and "bar", each of which has a non-shared "development" branch, the
super-module will track the submodule changes in branches called
"foo/development" and "bar/development".

Branches in abstract repositories are assumed to be shared and
conflict-free.

This commit will be reverted, as the assumptions are bogus!