Remove bogus error and assertion
It was firing for failures to map changed SVN directory paths to Git,
even when those directories were just being created and were empty.
Adding this to earlier reasons not to fail when unmapped SVN paths
were being deleted, we really have no business checking for this
condition at this point in the code.