Move <boost/detail/none_t.hpp> into optional.
It says in the file that it's part of optional.  This file should
actually be deleted - it's a near duplicate of <boost/none_t.hpp> (part
of optional) and doens't appear to be used anywhere, so I assume it was
an accident. But might as well put it in the right place.