Move the tools from sandbox/boost_docs/trunk to the correct repos.
I just noticed the 'sandbox/boost_docs' branch in the website repo,
which has nothing to do with the website. I've tried to move the
boostbook and quickbook parts of that branch to the correct repos.

I'm not sure what should be done with the rest, the 'doc' directory
appears to belong with the boost super-project. I don't think we really
need to preserve it.